Fixed a sorting error for t3, where the difference was being calculated at a lower precision than required. This caused extra bytes to be cast into int64_t, which led to incorrect time sorting.

+++ src/t3.c
@@ -75,7 +75,7 @@ int t3_compare(void const *a, void const *b) {
 	int64_t difference;
 
 	if ( ((t3_t *)a)->pulse == ((t3_t *)b)->pulse ) {
-		difference = ((t3_t *)a)->time - ((t3_t *)b)->time;
+		difference = (int64_t)((t3_t *)a)->time - (int64_t)((t3_t *)b)->time;
 	} else {
 		difference = ((t3_t *)a)->pulse - ((t3_t *)b)->pulse;
 	}
